requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
moviepy==1.0.3
openai-whisper==20231117
torch==2.1.1
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.2
pydantic-settings==2.1.0
ffmpeg-python==0.2.0
